Remove all mention of turn-on-artbollocks-mode
This function is no longer part of artbollocks-mode but the README and
the usage documentation in the source code both show it as the
preferred way to enable artbollocks-mode as a hook for other modes.
Attempting to use turn-on-artbollocks-mode can cause some obscure
errors, such as breaking log-edit-show-diff.  The correct way to
enable artbollocks-mode is to simply setup hooks to call that function
directly, and this patch changes the documentation to reflect that.

Signed-off-by: Eric James Michael Ritz <lobbyjones@gmail.com>