Improve optional parameter and region support
This changeset gracefully handles optional START and END parameters in
ARTBOLLOCKS-COUNT-WORDS and ARTBOLLOCKS-COUNT-SENTENCES.  When either
START or END are omitted, these functions will use POINT-MIN or
POINT-MAX, respectively.

I've also replaced several calls to (INTERACTIVE "r") to allow the user
to invoke the corresponding functions without specifying a region.  If a
region isn't specified, the entire buffer, subject to narrowing, is
used.